benchmarking: implement micro-VM suspend/resume suite phases 0–2
Promote ateom-microvm checkpoint/restore phase timings from logs to
ate.microvm.*.duration histograms (including OnDemand merge as its own
phase), and label atelet.snapshot.size with ate.sandbox.class so gVisor
and micro-VM compositions can be compared.

Widen glutton WriteDisk to int64, add DiskUser for rootfs-upper cycles
against the default glutton template, and register micro-VM baselines,
DurDir parity, rootfs size sweeps, and implicit-resume entries in
tests.yaml.

+++ internal/benchmarking/boomer/glutton/durdir_test.go
@@ -100,7 +100,7 @@ func TestDurDirUsesConfiguredFileSize(t *testing.T) {
 	if len(recorded) != 1 {
 		t.Fatalf("recorded write sizes: got %d calls, want 1", len(recorded))
 	}
-	if int64(recorded[0]) != configuredSize {
+	if recorded[0] != configuredSize {
 		t.Errorf("WriteDisk received size %d, want %d", recorded[0], configuredSize)
 	}
 }
